không cần setDefaultMark boi vi de bai mac dinh default mark của mỗi câu hỏi đều là 1 đổi mark từ int -> double bởi vì sau này có thể dùng để chia điểm câu hỏi nhiều đáp án.

+++ src/DataObjects/Question.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 public class Question implements Serializable {
     private String questionName;
     private String questionText;
-    private final float defaultMark = 1;
+    private final double defaultMark = 1.0;
     private ArrayList<Choice> choices = new ArrayList<Choice>();            // loai bo final
     private Answer ans;                                                 // them vao thuoc tinh Answer
 
@@ -26,7 +26,7 @@ public class Question implements Serializable {
         return questionName;
     }
 
-    public int getDefaultMark() {
+    public double getDefaultMark() {
         return defaultMark;
     }
